SoftwareDevelopment/EjercicioGuiado2/requirements.txt
astroid==2.14.1
dill==0.3.6
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.4.0
platformdirs==2.6.2
pylint==2.16.1
python-barcode==0.14.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
